ai-rag/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
openai
chromadb
unstructured
python-dotenv
